requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
astroid==2.6.6
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
distlib==0.3.2
filelock==3.0.12
identify==2.2.13
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.2.0
pre-commit==2.14.0
pylint==2.9.6
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
virtualenv==20.7.2
wrapt==1.12.1
yarl==1.6.3
